pyproject.toml
[project]
name = "manipulator-isaac-learning"
version = "0.1.0"
description = "Simulation-first manipulation learning with Isaac Lab + LeRobot"
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["numpy>=1.26.0", "python-dotenv>=1.0.0"]

[project.optional-dependencies]
sim = [
    "huggingface_hub>=0.23.0",
    "pyzmq>=26.0.0",
    "gymnasium>=0.29.0",
    "numpy>=1.26.0",
]
train = [
    "torch>=2.3.0",
    "torchvision>=0.18.0",
    "lerobot>=0.4.1",
    "pyzmq>=26.0.0",
    "gymnasium>=0.29.0",
    "numpy>=1.26.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
